Safari-611.1.10/LayoutTests/crypto/subtle/ecdsa-import-key-sign-p384-expected.txt

Test ECDSA signing operation with a P-384 private key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS signature.byteLength is 96
PASS successfullyParsed is true

TEST COMPLETE
